Parser: remove `Statement` entirely from the ast
